Add full functionality and CSS styles
Everything works for the most part now. There are still a few strange
bugs to sort out, like what happens when a user chooses one operator and
then another.

The main thing I might still need to add is the ability to use negative
numbers as the second number... right now the user cannot do that.

I think if I change how my code works, it might be better. Instead of
showing the operator in the display text and grabbing it from there, for
example, it might be better to have the operators switch the calculator
into a certain mode (like adding mode, etc.).